coding

+++ app/Http/Controllers/PostController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Blog;
+use App\Models\Comment;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -24,6 +25,11 @@ class PostController extends Controller
         $blog = array();
         if ($request->has('id')) {
             $blog = Blog::getBlogInfo($request->id);
+            if ($blog->created_user != Auth::id() and Auth::user()->role != 1) {
+                print_r($blog);die;
+                // can't update orther user's post if user is not admin
+                return redirect()->back();
+            }
         }
         return view('post-new-blog', ['blog' => $blog]);
     }
@@ -43,6 +49,12 @@ class PostController extends Controller
         ]);
         if ($data) {
             if ($request->has('id') and $request->id) {
+                if ($request->created_user != Auth::id() and Auth::user()->role != 1) {
+                    // can't update orther user's post if user is not admin
+                    return redirect()->back()->withErrors([
+                        'message' => "Can not modify other user's post.",
+                    ]);
+                }
                 Blog::updateBlog($data, $request->id);
             } else {
                 Blog::addBlog($data);
@@ -63,6 +75,13 @@ class PostController extends Controller
      */
     public function deleteBlog(Request $request)
     {
+        if ($request->created_user != Auth::id() and Auth::user()->role != 1) {
+            print_r($request->created_user);die;
+            // can't update orther user's post if user is not admin
+            return redirect()->back()->withErrors([
+                'message' => "Can not delete other user's post.",
+            ]);
+        }
         Blog::deleteBlog($request->id);
         return redirect()->back();
     }
@@ -78,7 +97,27 @@ class PostController extends Controller
     {
         $blog = Blog::getBlogInfo($request->id);
         if ($blog) {
-            return view('blog-detail', ['blog' => $blog]);
+            $comments = Comment::getListComment($blog->id);
+            return view('blog-detail', ['blog' => $blog, 'comments' => $comments]);
+        }
+        return redirect()->back();
+    }
+
+    /**
+     * addComment
+     *
+     * @param  mixed $request
+     * @return void
+     */
+    public function addComment(Request $request)
+    {
+        $data = $request->validate([
+            'blog_id' => 'required|numeric',
+            'content' => 'required'
+        ]);
+
+        if ($data) {
+            Comment::addComment($data);
         }
         return redirect()->back();
     }
